chore: move get_borsh_account and get_account to bench
