+++ governance-chat/program/tests/process_post_message.rs
@@ -13,5 +13,12 @@ async fn test_post_message() {
     let proposal_cookie = governance_chat_test.with_proposal().await;
 
     // Act
-    let _message_cookie = governance_chat_test.with_message(&proposal_cookie).await;
+    let message_cookie = governance_chat_test.with_message(&proposal_cookie).await;
+
+    // Assert
+    let message_data = governance_chat_test
+        .get_message_account(&message_cookie.address)
+        .await;
+
+    assert_eq!(message_data, message_cookie.account);
 }
